init new repo for ipldtool.
A better name may be forthcoming.  Naming is hard.

Some code has already incubated elsewhere (under the same licenses),
so a code drop will be coming shortly.
(A new repo is being made because I'm not having a lot of joy with
having more than one go module in a single git repo.)