extract the spec test common actions.
(Some of this might even be worth carrying all the way upstream into
the testmark testexec package?  Unclear, as yet.)

Also... fix it to actually compile.  And check if that worked.
It was broken when I rearranged some packages, and I didn't notice.
Ouch.  That was foolish, wasn't it, past self.