more error standardization and analyzability work.
Hoisted the function for the `schema parse` subcommand out and named it
so that the analyzer can look at it.

Started annotating it with the errors it returns.

This forced annotating more of the errors from other features it calls,
exactly as intended.  So, more wrapper functions are extracted,
which make sure whatever errors those functions return are tagged
with something we recognize explicitly before being returned further.
(This is borderline boilerplately... but honestly, once I got over it,
it really is making the main function here nice and linear, so maybe
these really did deserve functions anyway.)

The github.com/warpfork/go-ree/analyse tool is mostly happy with this.
The 'ipld.EncodeStreaming' call is still unhappy; gonna deal with that
in a subsequent commit.