begin put command.  skeletal.
A *lot* of work already went on upstream:
the go-ipld-prime library is introducing new storage APIs,
and there's a whole new "dsadapter" module,
which lets the new APIs be backed by the ancient go-datastore family.

We're starting to use _some_ of that in this commit already,
but this is also just a checkpoint commit about half way through.
We'll have this whole situation bundled up invisibly inside
the go-ipld-prime LinkSystem by the time we're done.

Also: a bunch of comments about configuring of storage.
Nothing final; those will turn into a real strategy shortly, one hopes.
(I'm going to try to let real UX drive this, as we play with
the CLI during its own development, via the testmark exec fixtures.)