workspace: begin introducing.
I want to use this to hold local storage,
and possibly some configuration.

The exact details may be subject to some evolution.
(You can see some of the earliest thoughts in comment form
in the put.go file -- in short, that putting all storage system
config in the CLI args every time feels unlikely to succeed,
so we need something more.)